Reorganiza a estrutura de endereço, promovendo a entidade a um módulo, para manter algumas responsabilidades decorrentes da segregaração na modelagem. Junto com a promoção, é feita criação da estrutura básica de módulo, com arquivos de teste, service e demais itens listados no commit.
