+++ src/address/address.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { AddressService } from './address.service';
+
+@Module({
+  providers: [AddressService],
+})
+export class AddressModule {}
