+++ src/address/address.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { AddressService } from './address.service';
+
+describe('AddressService', () => {
+  let service: AddressService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [AddressService],
+    }).compile();
+
+    service = module.get<AddressService>(AddressService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
